fix module_inception of daily_task
